+++ lib/attr_cipher.rb
@@ -1,3 +1,4 @@
 require 'attr_cipher/version'
+require 'attr_cipher/secret_exception'
 require 'attr_cipher/cipher'
 require 'attr_cipher/attr_cipher'