+++ lib/attr_cipher/secret_exception.rb
@@ -1,0 +1,4 @@
+module AttrCipher
+  class SecretException < ::StandardError
+  end
+end